tests: fix float vs int in backup test
(cherry picked from commit 7b0d2490a221cf589d4b5b85310522eb4efaf48b)

+++ qubes/tests/integ/backupcompatibility.py
@@ -149,7 +149,7 @@ class TC_00_BackupCompatibility(
         subprocess.check_call(["/usr/sbin/mkfs.ext4", "-q", "-F", filename])
 
     def create_volatile_img(self, filename):
-        self.create_sparse(filename, 11.5*2**30)
+        self.create_sparse(filename, int(11.5*2**30))
         # here used to be sfdisk call with "0,1024,S\n,10240,L\n" input,
         # but since sfdisk folks like to change command arguments in
         # incompatible way, have an partition table verbatim here
